refactor(platform): demote BillReminderScheduler core-base → core/platform
Domain-specific (banking/bills) — belongs in project layer, not framework.
core-base now exposes generic NotificationScheduler<T : NotificationRequest>
interface; BillReminderScheduler in core/platform/notification/bill/
implements with BillReminderSchedule : NotificationRequest.
