+++ core-base/platform/src/commonMain/kotlin/template/core/base/platform/di/PlatformModule.kt
@@ -11,7 +11,6 @@ package template.core.base.platform.di
 
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
-import org.koin.core.module.Module
 import org.koin.dsl.module
 import template.core.base.platform.garbage.GarbageCollectionManager
 import template.core.base.platform.garbage.GarbageCollectionManagerImpl
@@ -19,17 +18,4 @@ import template.core.base.platform.garbage.GarbageCollectionManagerImpl
 val platformModule = module {
     single<CoroutineDispatcher> { Dispatchers.Unconfined }
     single<GarbageCollectionManager> { GarbageCollectionManagerImpl(get()) }
-
-    includes(notificationModule)
 }
-
-/**
- * Platform-specific Koin bindings for the notification surface.
- *
- * Each target supplies its own [template.core.base.platform.notification.BillReminderScheduler]
- * actual — Android needs an `applicationContext`, iOS / desktop / JS / WasmJS need nothing — so
- * the binding is split into a per-platform [Module] rather than living in the shared
- * [platformModule]. Pulled in via `includes()` above so consumers only ever load
- * `platformModule`.
- */
-expect val notificationModule: Module
