--- core-base/platform/src/desktopMain/kotlin/template/core/base/platform/notification/BillReminderScheduler.desktop.kt
@@ -1,35 +1,0 @@
-/*
- * Copyright 2026 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
- */
-package template.core.base.platform.notification
-
-/**
- * Desktop (JVM) stub for [BillReminderScheduler].
- *
- * Compose Multiplatform on the desktop does not yet expose a portable system-notification
- * API across Windows/macOS/Linux, so this implementation deliberately no-ops. Bills are
- * still persisted by the caller, and the in-app "Upcoming Bills" surface acts as the
- * fallback reminder when the user opens the app.
- *
- * Forks may swap this for an `awt.SystemTray` / `java.awt.TrayIcon` implementation, or
- * a Windows toast / macOS user-notification bridge, without touching common code.
- */
-actual class BillReminderScheduler {
-    actual suspend fun schedule(bill: BillReminderSchedule) {
-        // Intentional no-op — see KDoc above.
-    }
-
-    actual suspend fun cancel(billId: String) {
-        // No persisted state on desktop, so nothing to cancel.
-    }
-
-    actual suspend fun cancelAll() {
-        // No persisted state on desktop, so nothing to cancel.
-    }
-}
