--- core-base/platform/src/jsMain/kotlin/template/core/base/platform/notification/BillReminderScheduler.js.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright 2026 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
- */
-package template.core.base.platform.notification
-
-/**
- * Kotlin/JS stub for [BillReminderScheduler].
- *
- * The browser `Notification` API only fires while the tab is open and permission has been
- * granted; out-of-band scheduling requires a Service Worker, which the template doesn't
- * yet wire up. This stub no-ops so common code compiles cleanly — bills still persist
- * and the in-app banner fires on next open.
- */
-actual class BillReminderScheduler {
-    actual suspend fun schedule(bill: BillReminderSchedule) {
-        // Intentional no-op until Service Worker integration ships.
-    }
-
-    actual suspend fun cancel(billId: String) {
-        // No persisted state on web, so nothing to cancel.
-    }
-
-    actual suspend fun cancelAll() {
-        // No persisted state on web, so nothing to cancel.
-    }
-}
