+++ feature/bills/build.gradle.kts
@@ -23,6 +23,7 @@ kotlin {
             implementation(projects.core.ui)
             implementation(projects.core.data)
             implementation(projects.core.model)
+            implementation(projects.core.platform)
             implementation(projects.coreBase.store)
             implementation(projects.coreBase.platform)
             implementation(projects.coreBase.ui)
